MOVET: CoreData func CoreDataManager파일로 이동
