+++ MODA/Controllers/SignUpViewController.swift
@@ -215,10 +215,6 @@ class SignUpViewController: UIViewController {
     @objc func saveData() {
         // 데이터 저장.
         print("save Data")
-        // 현재 시간
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd HH:mm:ss"
-        let currentDate = formatter.string(from: Date())
         
         if nameTextField.text == "" {
             alertMessage(message: "name")
@@ -230,173 +226,7 @@ class SignUpViewController: UIViewController {
             alertMessage(message: "pw check")
         } else {
             let data = UsersInfo(name: nameTextField.text ?? "", id: idTextField.text ?? "", pw: pwTextField.text ?? "", email: emailTextField.text ?? "")
-            insertCoreData(userInfo: data)
-        }
-    }
-    
-    // MARK: Core Data 함수
-    // Core Data Insert
-    func insertCoreData(userInfo: UsersInfo) -> Bool {
-        print("insert Core Data")
-        print("userinfo : \(userInfo)")
-        // NSManagedObjectContext를 가져온다.
-        let appDelegate = UIApplication.shared.delegate as! AppDelegate
-        let context = appDelegate.persistentContainer.viewContext
-        // Entity를 가져온다.
-        let entity = NSEntityDescription.entity(forEntityName: "UserInfo", in: context)
-        
-        if let entity = entity {
-            let managedObject = NSManagedObject(entity: entity, insertInto: context)
-            
-            managedObject.setValue(userInfo.name, forKey: "name")
-            managedObject.setValue(userInfo.id, forKey: "id")
-            managedObject.setValue(userInfo.pw, forKey: "pw")
-            managedObject.setValue(userInfo.email, forKey: "email")
-            
-            do {
-                try context.save()
-                return true
-            } catch {
-                print(error.localizedDescription)
-                return false
-            }
-        } else {
-            return false
-        }
-    }
-    
-    // Core Data Count
-    func countCoreData<T: NSManagedObject>(request: NSFetchRequest<T>) -> Int? {
-        print("count Core Data")
-        // NSManagedObjectContext를 가져온다.
-        let appDelegate = UIApplication.shared.delegate as! AppDelegate
-        let context = appDelegate.persistentContainer.viewContext
-        do {
-            let count = try context.count(for: request)
-            return count
-        } catch {
-            return nil
-        }
-    }
-    
-    // Core Data Fetch
-    func fetchCoreData<T: NSManagedObject>(request: NSFetchRequest<T>) -> [T] {
-        print("fetch Core Data")
-        // NSManagedObjectContext를 가져온다.
-        let appDelegate = UIApplication.shared.delegate as! AppDelegate
-        let context = appDelegate.persistentContainer.viewContext
-        
-        do {
-            let fetchResult = try context.fetch(request)
-            return fetchResult
-        } catch {
-            print(error.localizedDescription)
-            return []
-        }
-    }
-    
-    // ID가 동일한 Core Data Fetch
-    func fetchIdCoreData<T: NSManagedObject>(request: NSFetchRequest<T>, id: String) -> [T] {
-        print("fetch Core Data")
-        // NSManagedObjectContext를 가져온다.
-        let appDelegate = UIApplication.shared.delegate as! AppDelegate
-        let context = appDelegate.persistentContainer.viewContext
-        
-        request.predicate = NSPredicate(format: "id = %@", id)
-        
-        do {
-            let fetchResult = try context.fetch(request)
-            return fetchResult
-        } catch {
-            print(error.localizedDescription)
-            return []
-        }
-    }
-    
-    // All Core Data Delete
-    func deleteAllCoreData<T: NSManagedObject>(request: NSFetchRequest<T>) -> Bool {
-        print("delete All Core Data")
-        // NSManagedObjectContext를 가져온다.
-        let appDelegate = UIApplication.shared.delegate as! AppDelegate
-        let context = appDelegate.persistentContainer.viewContext
-        
-        let request: NSFetchRequest<NSFetchRequestResult> = T.fetchRequest()
-        let delete = NSBatchDeleteRequest(fetchRequest: request)
-        do {
-            try context.execute(delete)
-            return true
-        } catch {
-            return false
-        }
-    }
-    
-    // Selected Core Data Delete
-    func deleteSelectedCoreData(object: NSManagedObject) -> Bool {
-        print("delete selected Core Data")
-        // NSManagedObjectContext를 가져온다.
-        let appDelegate = UIApplication.shared.delegate as! AppDelegate
-        let context = appDelegate.persistentContainer.viewContext
-        
-        context.delete(object)
-        do {
-            try context.save()
-            return true
-        } catch {
-            return false
-        }
-    }
-    
-    // My List Core Data Delete - 임의로 생성해 놓은 함수
-    // 해당되는 id 삭제
-    // 삭제를 하고 새로이 저장을 해야 해당되는 내용이 저장됨.
-    func deleteMyListCoreData<T: NSManagedObject>(request: NSFetchRequest<T>, id: String) -> Bool {
-        let appDelegate = UIApplication.shared.delegate as! AppDelegate
-        let context = appDelegate.persistentContainer.viewContext
-        
-        request.predicate = NSPredicate(format: "id = %@", id)
-        
-        do {
-            let fetchResult = try context.fetch(request)
-            if fetchResult.count == 0 {
-                print("deleteMyListCoreData : false")
-                return false
-            } else {
-                context.delete(fetchResult[0])
-                try context.save()
-                print("deleteMyListCoreData : true")
-                return true
-            }
-        } catch {
-            print(error.localizedDescription)
-            return false
-        }
-    }
-    
-    // 사용자 정보 수정
-    func editUserInfoCoreData<T: NSManagedObject>(request: NSFetchRequest<T>, id: String, name: String) -> Bool {
-        let appDelegate = UIApplication.shared.delegate as! AppDelegate
-        let context = appDelegate.persistentContainer.viewContext
-        
-        request.predicate = NSPredicate(format: "id = %@", id)
-        print("request : \(request)")
-        do {
-            let fetchResult = try context.fetch(request)
-            print("fetchResult : \(fetchResult)")
-            if fetchResult.count == 0 {
-                print("editUserInfoCoreData : false")
-                return false
-            } else {
-                // 수정.
-                for result in fetchResult {
-                    result.setValue(name, forKey: "name")
-                }
-                try context.save()
-                print("editUserInfoCoreData : true")
-                return true
-            }
-        } catch {
-            print(error.localizedDescription)
-            return false
+            CoreDataManager.shared.insertCoreData(userInfo: data)
         }
     }
     
@@ -414,7 +244,7 @@ class SignUpViewController: UIViewController {
     @objc func fetchData() {
         print("fetch button")
         let coreData = UserInfo.fetchRequest()
-        let fetch = fetchCoreData(request: coreData)
+        let fetch = CoreDataManager.shared.fetchCoreData(request: coreData)
         // 데이터 하나씩 출력
         for result in fetch {
             let data = UsersInfo(name: result.name ?? "", id: result.id ?? "", pw: result.pw ?? "", email: result.email ?? "")
@@ -437,7 +267,7 @@ class SignUpViewController: UIViewController {
         print("delete All Button")
         let coreData = UserInfo.fetchRequest()
 //        deleteAllCoreData(request: coreData)
-        editUserInfoCoreData(request: coreData, id: idTextField.text ?? "", name: nameTextField.text ?? "")
+        CoreDataManager.shared.editUserInfoCoreData(request: coreData, id: idTextField.text ?? "", name: nameTextField.text ?? "")
     }
     
     private lazy var oneDeleteButton: UIButton = {
@@ -452,11 +282,11 @@ class SignUpViewController: UIViewController {
     @objc func oneDeleteData() {
         print("One delete Button")
         let coreData = UserInfo.fetchRequest()
-        let fetchResult = fetchCoreData(request: UserInfo.fetchRequest())
+        let fetchResult = CoreDataManager.shared.fetchCoreData(request: UserInfo.fetchRequest())
         
-        let count = countCoreData(request: UserInfo.fetchRequest())
+        let count = CoreDataManager.shared.countCoreData(request: UserInfo.fetchRequest())
         print("Core 데이터 Count : \(count)")
-        deleteMyListCoreData(request: coreData, id: idTextField.text ?? "")
+        CoreDataManager.shared.deleteMyListCoreData(request: coreData, id: idTextField.text ?? "")
     }
     
     func alertMessage(message: String) {
